fix: 开放题上次分数和次数

+++ backend/internal/app/model/response/challenge.go
@@ -52,7 +52,7 @@ type GetUserOpenQuestDetailListV2 struct {
 	UserScore             int64          `gorm:"column:user_score" form:"user_score" json:"user_score"`    // 用户分数
 	NickName              *string        `gorm:"column:nickname;type:varchar(200);default:''" json:"nickname" form:"nickname"`
 	SubmitCount           int64          `gorm:"column:submit_count" json:"submit_count"`
-	LastScore             int64          `gorm:"column:last_score" json:"last_score"`
+	LastScore             float64        `gorm:"column:last_score" json:"last_score"`
 }
 
 type GetUserQuestDetailResponse struct {
